todo: perf

+++ shaders/lib/buffers/addr.glsl
@@ -5,10 +5,10 @@
 #include "/lib/common/tiled_addr.glsl"
 
 // ===========================================================================
-// Vec4-based SSBO addressing — tiledAddr8x8 for 8×8 tile-encoded SSBO access
+// Uvec4-based SSBO addressing — tiledAddr8x8 for 8×8 tile-encoded SSBO access
 // ===========================================================================
-// All SSBOs are stored as vec4 data[]. Each abstract "image" is a resolution×resolution
-// vec4 grid encoded in 8×8 tiles. N selects the abstract image layer.
+// Screen SSBOs use raw uvec4 words. Packed values stay integer end-to-end;
+// genuine FP32 lanes cross the storage boundary through vector bitcasts.
 
 uint addr(uint N, uvec2 xy) {
     // Clamp to resolution_global to prevent out-of-bounds coordinates from corrupting tile calculation
@@ -21,11 +21,11 @@ uint addr(uint N, ivec2 xy) {
 }
 
 // ===========================================================================
-// SSBO declarations — 4 bindings, all vec4 data[]
+// SSBO declarations — screen virtual layers use uvec4 data[]
 // ===========================================================================
 
 layout(std430, set = 3, binding = 0) buffer GeometryMaterialBuffer {
-    vec4 data[];
+    uvec4 data[];
 } geomBuffer;
 
 layout(std430, set = 3, binding = 2) buffer DiffuseBuffer {
@@ -33,11 +33,11 @@ layout(std430, set = 3, binding = 2) buffer DiffuseBuffer {
 } diffuseBuffer;
 
 layout(std430, set = 3, binding = 3) buffer ReflectBuffer {
-    vec4 data[];
+    uvec4 data[];
 } reflectBuffer;
 
 layout(std430, set = 3, binding = 4) buffer RefractBuffer {
-    vec4 data[];
+    uvec4 data[];
 } refractBuffer;
 
 layout(std430, set = 3, binding = 5) buffer RadianceCacheBuffer {
